Moved scenario, registers to modelrun from sosmodel
